+++ src/core/components/ReactiveInputDialog.tsx
@@ -2,10 +2,10 @@
 // drag reordering https://www.npmjs.com/package/react-smooth-dnd
 // https://codesandbox.io/s/material-ui-sortable-list-with-react-smooth-dnd-swrqx?file=/src/index.js:113-129
 
-import { Button, Dialog, DialogContent, DialogTitle, useMediaQuery } from "@mui/material";
-import { useTheme } from "@mui/material/styles";
+import { Button, DialogContent, DialogTitle } from "@mui/material";
 import React, { Suspense } from "react";
 import { CMDialogContentText, DialogActionsCM } from "./CMCoreComponents2";
+import { ResponsiveDialog } from "./ResponsiveDialog";
 
 
 ////////////////////////////////////////////////////////////////
@@ -20,9 +20,6 @@ export interface ReactiveInputDialogProps {
     defaultAction?: () => void; // this is the default action when Enter is pressed
 };
 export const ReactiveInputDialog = ({ open = true, ...props }: React.PropsWithChildren<ReactiveInputDialogProps>) => {
-    const theme = useTheme();
-    const fullScreen = useMediaQuery(theme.breakpoints.down('md'));
-
     const handleKeyDown = (event: React.KeyboardEvent<HTMLDivElement>) => {
         props.onKeyDown?.(event);
         if (event.isPropagationStopped()) {
@@ -43,20 +40,19 @@ export const ReactiveInputDialog = ({ open = true, ...props }: React.PropsWithCh
     };
 
     return (
-        <Dialog
-            className={`ReactiveInputDialog ${props.className} ${fullScreen ? "smallScreen" : "bigScreen"}`}
+        <ResponsiveDialog
+            className={`ReactiveInputDialog ${props.className ?? ""}`}
             open={open}
             style={props.style}
             onClose={props.onCancel}
             scroll="paper"
-            fullScreen={fullScreen}
             disableRestoreFocus={true} // this is required to allow the autofocus work on buttons. https://stackoverflow.com/questions/75644447/autofocus-not-working-on-open-form-dialog-with-button-component-in-material-ui-v
             onKeyDown={handleKeyDown}
         >
             <Suspense>
                 {props.children}
             </Suspense>
-        </Dialog>
+        </ResponsiveDialog>
     );
 };